Move all tests added to serialize_program.cpp to a separate file
