+++ test/serialize_program.cpp
@@ -28,7 +28,6 @@
 #include <migraphx/make_op.hpp>
 #include <migraphx/instruction.hpp>
 #include <migraphx/iterator_for.hpp>
-#include <migraphx/stringutils.hpp>
 
 #include <cstdio>
 
@@ -246,61 +245,6 @@ TEST_CASE(no_debug_symbols_roundtrip)
 
     auto syms2 = collect_debug_symbols(*p2.get_main_module());
     EXPECT(syms2.empty());
-}
-
-static migraphx::program create_program_with_comment()
-{
-    migraphx::program p;
-    auto* mm = p.get_main_module();
-    mm->add_instruction(migraphx::make_op("@comment", {{"text", "gpu::mlir_op test"}}), {});
-    auto x   = mm->add_parameter("x", {migraphx::shape::int32_type});
-    auto two = mm->add_literal(2);
-    auto add = mm->add_instruction(migraphx::make_op("add"), x, two);
-    mm->add_return({add});
-    return p;
-}
-
-TEST_CASE(comment_eval)
-{
-    migraphx::program p = create_program_with_comment();
-    auto result         = p.eval({{"x", migraphx::literal{3}.get_argument()}}).back();
-    EXPECT(result == migraphx::literal{5});
-}
-
-TEST_CASE(comment_as_value)
-{
-    migraphx::program p1 = create_program_with_comment();
-    migraphx::program p2;
-    p2.from_value(p1.to_value());
-    EXPECT(p1.sort() == p2.sort());
-}
-
-TEST_CASE(comment_as_msgpack)
-{
-    migraphx::file_options options;
-    options.format           = "msgpack";
-    migraphx::program p1     = create_program_with_comment();
-    std::vector<char> buffer = migraphx::save_buffer(p1, options);
-    migraphx::program p2     = migraphx::load_buffer(buffer, options);
-    EXPECT(p1.sort() == p2.sort());
-}
-
-TEST_CASE(comment_as_file)
-{
-    std::string filename = "migraphx_comment_program.mxr";
-    migraphx::program p1 = create_program_with_comment();
-    migraphx::save(p1, filename);
-    migraphx::program p2 = migraphx::load(filename);
-    std::remove(filename.c_str());
-    EXPECT(p1.sort() == p2.sort());
-}
-
-TEST_CASE(comment_print)
-{
-    migraphx::program p = create_program_with_comment();
-    auto s              = migraphx::to_string(p);
-    EXPECT(s.find("@comment") != std::string::npos);
-    EXPECT(s.find("gpu::mlir_op test") != std::string::npos);
 }
 
 int main(int argc, const char* argv[]) { test::run(argc, argv); }
